feat(agent): added Windows support
