+++ Dockerfile.windows
@@ -1,0 +1,9 @@
+FROM microsoft/nanoserver
+
+USER ContainerAdministrator
+
+WORKDIR /app
+
+COPY dist/agent.exe /app/
+
+ENTRYPOINT ["C:/app/agent.exe"]
