+++ build.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+ARCHIVE_BUILD_FOLDER="/tmp/portainer-builds"
+
 if [[ $# -ne 1 ]] ; then
   echo "Usage: $(basename $0) <VERSION>"
   exit 1
@@ -18,9 +20,20 @@ function build_and_push_image() {
   docker push "portainer/agent:${1}"
 }
 
+function build_archive() {
+  BUILD_FOLDER="${ARCHIVE_BUILD_FOLDER}/$1"
+  rm -rf ${BUILD_FOLDER} && mkdir -pv ${BUILD_FOLDER}/agent
+  mv dist/* ${BUILD_FOLDER}/agent/
+  cd ${BUILD_FOLDER}
+  tar cvpfz "portainer-agent-${VERSION}-$1.tar.gz" agent
+  mv "portainer-agent-${VERSION}-$1.tar.gz" ${ARCHIVE_BUILD_FOLDER}/
+  cd -
+}
+
 function build_binary() {
+  name="agent"; if [ "$(echo "$tag" | cut -c1)"  = "w" ]; then name="${name}.exe"; fi
   docker run --rm -tv $(pwd):/src -e BUILD_GOOS="$1" -e BUILD_GOARCH="$2" portainer/golang-builder:cross-platform /src/cmd/agent
-  mv "cmd/agent/agent-${1}-${2}" dist/agent
+  mv "cmd/agent/agent-${1}-${2}" dist/${name}
 }
 
 function build_all() {
@@ -29,11 +42,16 @@ function build_all() {
     arch=`echo ${tag} | cut -d \- -f 2`
 
     build_binary "${os}" "${arch}"
-    build_and_push_image "${tag}"
+    if [ `echo $tag | cut -d \- -f 1` == 'linux' ]
+    then
+      build_and_push_image "${tag}"
+    else
+      build_archive "$tag"
+    fi
     clean
   done
 }
 
-build_all 'linux-amd64 linux-arm linux-arm64 linux-ppc64le linux-s390x'
+build_all 'linux-amd64 linux-arm linux-arm64 linux-ppc64le linux-s390x windows-amd64'
 
 exit 0
