+++ constants/constants_windows.go
@@ -1,0 +1,8 @@
+// +build windows
+
+package constants
+
+const (
+	// SystemVolumePath represents the path where the volumes are stored on the filesystem
+	SystemVolumePath = `C:\ProgramData\Docker\volumes`
+)
