+++ http/handler/websocket/websocket_linux.go
@@ -1,0 +1,11 @@
+// +build linux
+
+package websocket
+
+import (
+	"net"
+)
+
+func createDial() (net.Conn, error) {
+	return net.Dial("unix", "/var/run/docker.sock")
+}
