+++ http/handler/websocket/websocket_windows.go
@@ -1,0 +1,13 @@
+// +build windows
+
+package websocket
+
+import (
+	"net"
+
+	"github.com/Microsoft/go-winio"
+)
+
+func createDial() (net.Conn, error) {
+	return winio.DialPipe("//./pipe/docker_engine", nil)
+}
