+++ http/proxy/local_linux.go
@@ -1,0 +1,29 @@
+// +build linux
+
+package proxy
+
+import (
+	"log"
+	"net"
+	"net/http"
+	"os"
+
+	"bitbucket.org/portainer/agent"
+)
+
+// NewLocalProxy returns a pointer to a LocalProxy.
+func NewLocalProxy(clusterService agent.ClusterService) *LocalProxy {
+	proxy := &LocalProxy{
+		transport: newSocketTransport("/var/run/docker.sock"),
+		logger:    log.New(os.Stderr, "", log.LstdFlags),
+	}
+	return proxy
+}
+
+func newSocketTransport(socketPath string) *http.Transport {
+	return &http.Transport{
+		Dial: func(proto, addr string) (conn net.Conn, err error) {
+			return net.Dial("unix", socketPath)
+		},
+	}
+}
